Colorful Card Game
There is a card array with nine slots.
Each slot includes a letter between A and D (A, B, C, D) in one of the three colors (red, blue, and green).
The array contain 3 cards in red color, 3 cards in blue color, and 3 cards in green color.
The player gets some score, depending on the letters and their colors that land along with the array.
If the consecutive three slots show the same letter in the same color, he/she gets 33 points
If the consecutive three slots show the same letter in different colors, he/she gets 28 points
If the consecutive three slots show the same letter, but not all the same or different color, he/she gets 22 points
If the consecutive three slots show consecutive letters in the same color, he/she gets 18 points
If the consecutive three slots show consecutive letters in different colors, he/she gets 16 points
If the consecutive three slots show consecutive letters, but not all the same or different color, he/she gets 14 points
If the consecutive three slots show the same color, but not all the same letter and not consecutive, he/she gets 12 points
If the consecutive three slots show different colors, but not all the same letter and not consecutive, he/she gets 10 points
Otherwise, he/she gets 0 points

Program randomly generates the array and then prints the score the player gets.
After that, the program ask the user whether he/she wants to play again or not. According to the choice of the user, the game will be repeated or over.

The high-score-table contains the top-10 game scores of all time.
The high-score-table is readead from the text file and inserted into the high-score array(s).